Split the big map classes and conversion to a separate package
 * Reduce the clutter of the database related classes
